Fix the calculation of numbers of samples to read.
On branch master
Your branch is up to date with 'origin/master'.

Changes to be committed:
	modified:   src/MAX30100.cpp

The calculation of numbers of samples to read in MAX30100::readFifoData() doesn't consider the situation when the FIFO is full. I add the check for whether the FIFO overflow counter is not 0 to detect this. If the FIFO is full, set the number of samples to read to 16. This can prevent not reading any data because the FIFO write pointer pointed to the same position as FIFO read pointer does.